Factor out some macro definitions into an example system we can use
